+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## [Unreleased]
 ### Added
+- `blank` argument for `Str` attribute - specifies whether value can be blank or not, defaults to `False`
 ### Fixed
 - proxy schema needs to be able to apply defaults (#10)
 - issue where list validation would sometimes lead to duplicate items added to the list
 ### Changed
+- `Directory` attribute will forfeit validation if `blank` is `True` and value is `""`
 ### Deprecated
 ### Removed
 ### Security